Extend FilesPipeline instead of using custom one
 * Removed file.py whicih had custom implementation on pipeline
 * Now extending FilesPipeline
