+++ opengazettes/opengazettes/settings.py
@@ -66,7 +66,7 @@ DOWNLOAD_DELAY = 3
 # Configure item pipelines
 # See http://scrapy.readthedocs.org/en/latest/topics/item-pipeline.html
 ITEM_PIPELINES = {
-   'opengazettes.files.FilesPipeline': True,
+   'opengazettes.pipelines.OpengazettesFilesPipeline': True,
 }
 
 # Enable and configure the AutoThrottle extension (disabled by default)
